subfeat: comments last update , added some validation

+++ src/modules/comments/comments.service.ts
@@ -1,54 +1,94 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from '../../prisma.service';
 import { CommentDto } from '../../common/dto/comment.dto';
 import { CreateCommentInput } from './dto/createComment.input';
 import { UpdateCommentInput } from './dto/updateComment.input';
-
 import { validate } from 'class-validator';
 
 @Injectable()
 export class CommentsService {
   constructor(private prisma: PrismaService) {}
 
+  /**
+   * Fetch all comments.
+   */
   async findAll(): Promise<CommentDto[]> {
     return this.prisma.comment.findMany();
   }
 
+  /**
+   * Fetch a single comment by ID.
+   * @param id - The ID of the comment to fetch.
+   */
   async findOne(id: number): Promise<CommentDto> {
-    return this.prisma.comment.findUnique({
+    const comment = await this.prisma.comment.findUnique({
       where: { id },
     });
+
+    if (!comment) {
+      throw new Error('Comment not found');
+    }
+
+    return comment;
   }
 
   async create(
     createCommentInput: CreateCommentInput,
     userId: number,
   ): Promise<CommentDto> {
-    const errors = await validate(createCommentInput);
-    if (errors.length > 0) {
-      throw new Error('Validation failed');
+    // Check if the post exists
+    const post = await this.prisma.post.findUnique({
+      where: { id: createCommentInput.postId },
+    });
+
+    if (!post) {
+      throw new NotFoundException('Post not found');
     }
 
     return this.prisma.comment.create({
       data: {
-        ...createCommentInput,
+        content: createCommentInput.content,
+        postId: createCommentInput.postId,
         userId,
       },
     });
   }
 
   async update(
-    id: number,
     updateCommentInput: UpdateCommentInput,
     userId: number,
   ): Promise<CommentDto> {
-    const errors = await validate(updateCommentInput);
-    if (errors.length > 0) {
-      throw new Error('Validation failed');
+    // search for comment and post
+    const { commentId } = updateCommentInput;
+    const [post, comment] = await Promise.all([
+      this.prisma.post.findUnique({
+        where: { id: updateCommentInput.postId },
+      }),
+      this.prisma.comment.findUnique({
+        where: { id: commentId },
+      }),
+    ]);
+
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+
+    if (!comment) {
+      throw new NotFoundException('Comment not found');
+    }
+
+    if (comment.userId !== userId) {
+      throw new ForbiddenException(
+        'You are not authorized to update this comment',
+      );
     }
 
     return this.prisma.comment.update({
-      where: { id, userId },
+      where: { id: commentId },
       data: {
         content: updateCommentInput.content,
         isEdited: true,
@@ -56,14 +96,28 @@ export class CommentsService {
     });
   }
 
-  async remove(commentId: number, userId: number): Promise<boolean> {
+  async remove(commentId: number, userId: number): Promise<string> {
+    const comment = await this.prisma.comment.findUnique({
+      where: { id: commentId },
+    });
+
+    if (!comment) {
+      throw new NotFoundException('Comment not found');
+    }
+
+    if (comment.userId !== userId) {
+      throw new ForbiddenException(
+        'You are not authorized to update this comment',
+      );
+    }
+
     try {
       await this.prisma.comment.delete({
-        where: { id: commentId, userId },
+        where: { id: commentId },
       });
-      return true;
-    } catch {
-      return false;
+      return 'comment deleted succsufyl';
+    } catch (error) {
+      throw new error();
     }
   }
 }
